fix: remove import infinite loop
